issue 296 extract interface and renaming to XslReport
