+++ src/main/java/org/jpeek/App.java
@@ -177,7 +177,7 @@ public final class App {
         final Collection<Report> reports = new LinkedList<>();
         if (this.params.containsKey("LCOM")) {
             reports.add(
-                new Report(
+                new XslReport(
                     chain.transform(skeleton),
                     "LCOM", this.params, 10.0d, -5.0d
                 )
@@ -185,7 +185,7 @@ public final class App {
         }
         if (this.params.containsKey("CAMC")) {
             reports.add(
-                new Report(
+                new XslReport(
                     chain.transform(skeleton),
                     "CAMC", this.params
                 )
@@ -193,7 +193,7 @@ public final class App {
         }
         if (this.params.containsKey("MMAC")) {
             reports.add(
-                new Report(
+                new XslReport(
                     chain.transform(skeleton),
                     "MMAC", this.params, 0.5d, 0.1d
                 )
@@ -201,7 +201,7 @@ public final class App {
         }
         if (this.params.containsKey("LCOM5")) {
             reports.add(
-                new Report(
+                new XslReport(
                     chain.transform(skeleton),
                     "LCOM5", this.params, 0.5d, -0.1d
                 )
@@ -209,7 +209,7 @@ public final class App {
         }
         if (this.params.containsKey("NHD")) {
             reports.add(
-                new Report(
+                new XslReport(
                     chain.transform(skeleton),
                     "NHD"
                 )
@@ -217,7 +217,7 @@ public final class App {
         }
         if (this.params.containsKey("LCOM2")) {
             reports.add(
-                new Report(
+                new XslReport(
                     chain.transform(skeleton),
                     "LCOM2", this.params
                 )
@@ -225,7 +225,7 @@ public final class App {
         }
         if (this.params.containsKey("LCOM3")) {
             reports.add(
-                new Report(
+                new XslReport(
                     chain.transform(skeleton),
                     "LCOM3", this.params
                 )
@@ -233,7 +233,7 @@ public final class App {
         }
         if (this.params.containsKey("SCOM")) {
             reports.add(
-                new Report(
+                new XslReport(
                     chain.transform(skeleton),
                     "SCOM", this.params
                 )
@@ -241,7 +241,7 @@ public final class App {
         }
         if (this.params.containsKey("OCC")) {
             reports.add(
-                new Report(
+                new XslReport(
                     chain.transform(skeleton),
                     "OCC", this.params
                 )
@@ -249,7 +249,7 @@ public final class App {
         }
         if (this.params.containsKey("PCC")) {
             reports.add(
-                new Report(
+                new XslReport(
                     chain.transform(skeleton),
                     "PCC"
                 )
@@ -257,7 +257,7 @@ public final class App {
         }
         if (this.params.containsKey("TCC")) {
             reports.add(
-                new Report(
+                new XslReport(
                     chain.transform(skeleton),
                     "TCC"
                 )
@@ -265,7 +265,7 @@ public final class App {
         }
         if (this.params.containsKey("LCC")) {
             reports.add(
-                new Report(
+                new XslReport(
                     chain.transform(skeleton),
                     "LCC"
                 )
@@ -273,7 +273,7 @@ public final class App {
         }
         if (this.params.containsKey("CCM")) {
             reports.add(
-                new Report(
+                new XslReport(
                     chain.transform(skeleton),
                     "CCM"
                 )
@@ -281,7 +281,7 @@ public final class App {
         }
         if (this.params.containsKey("MWE")) {
             reports.add(
-                new Report(
+                new XslReport(
                     chain.transform(skeleton),
                     "MWE"
                 )
