+++ src/test/java/org/jpeek/XslReportTest.java
@@ -37,7 +37,7 @@ import org.xembly.Directives;
 import org.xembly.Xembler;
 
 /**
- * Test case for {@link Report}.
+ * Test case for {@link XslReport}.
  * @author Yegor Bugayenko (yegor256@gmail.com)
  * @version $Id$
  * @since 0.4
@@ -46,12 +46,12 @@ import org.xembly.Xembler;
  * @checkstyle JavadocTagsCheck (500 lines)
  */
 @SuppressWarnings("PMD.AvoidDuplicateLiterals")
-public final class ReportTest {
+public final class XslReportTest {
 
     @Test
     public void createsXmlReport() throws IOException {
         final Path output = Files.createTempDirectory("");
-        new Report(new Skeleton(new FakeBase()).xml(), "LCOM").save(output);
+        new XslReport(new Skeleton(new FakeBase()).xml(), "LCOM").save(output);
         new Assertion<>(
             "Must LCOM.xml file exists",
             Files.exists(output.resolve("LCOM.xml")),
@@ -67,7 +67,7 @@ public final class ReportTest {
     @Test
     public void createsXmlReportWithXpaths() throws IOException {
         final Path output = Files.createTempDirectory("");
-        new Report(
+        new XslReport(
             new Skeleton(
                 new FakeBase(
                     "NoMethods", "Bar", "OverloadMethods",
@@ -92,7 +92,7 @@ public final class ReportTest {
     @Test
     public void createsXmlReportWithEmptyProject() throws IOException {
         final Path output = Files.createTempDirectory("");
-        new Report(new Skeleton(new FakeBase()).xml(), "LCOM").save(output);
+        new XslReport(new Skeleton(new FakeBase()).xml(), "LCOM").save(output);
         new Assertion<>(
             "Report for empty project created",
             XhtmlMatchers.xhtml(
@@ -107,7 +107,7 @@ public final class ReportTest {
     @Test
     public void createsFullXmlReport() throws IOException {
         final Path output = Files.createTempDirectory("");
-        new Report(
+        new XslReport(
             new XMLDocument(
                 new Xembler(
                     new Directives()
@@ -144,7 +144,7 @@ public final class ReportTest {
     @Test
     public void setsCorrectSchemaLocation() throws IOException {
         final Path output = Files.createTempDirectory("");
-        new Report(new Skeleton(new FakeBase()).xml(), "LCOM").save(output);
+        new XslReport(new Skeleton(new FakeBase()).xml(), "LCOM").save(output);
         new Assertion<>(
             "Must have correct schema location",
             XhtmlMatchers.xhtml(
